# library-manager-backend/src/main/java/com/igriss/library_manager/authentication/AuthenticationResponse.java
package com.igriss.library_manager.authentication;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}