# library-manager-backend/src/main/java/com/igriss/library_manager/book/BorrowedBookResponse.java
package com.igriss.library_manager.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Long id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}