# library-manager-backend/src/main/java/com/igriss/library_manager/feedback/FeedbackResponse.java
package com.igriss.library_manager.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}